backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-spectacular==0.27.0
python-telegram-bot==20.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
